TEMP: breaking changes to enable sampling from pretrained policy
We always train gaussian policies with RL (because could not get
discrete working previously), but we expect datasets to save actions
as one-hot vectors. In other words, we always collect data from env
with discrete=True. To accommodate this with sampling from a trained
policy, we transform the action from continuous to discrete in the
sampling procedure.